Removed WorkerThreadContext class
