+++ es6/node-worker.js
@@ -7,6 +7,8 @@
  * under the MIT License. See LICENSE file for details.
  */
 
-module.exports = function() {
+"use strict";
 
-}
+const { parentPort, workerData } = require("node:worker_threads");
+
+module.exports = require("./worker-script");
