--- es6/webcalls-num.js
@@ -1,50 +1,0 @@
-module.exports = {
-  WEB_KERNEL_MEMORY_ID: 1,
-  WEB_KERNEL_MODULE_ID: 2,
-  WEI_FALSE_BOOLEAN_ID: 6,
-  WEI_TRUE_BOOLEAN_ID: 7,
-  WEB_NULL_OBJECT_ID: 8,
-  WEB_GLOBAL_THIS_ID: 9,
-  WEB_SCRIPT_URL_ID: 10,
-  
-  WEB_ConsoleWrite: 11,
-  WEB_ObjectRelease: 12,
-  WEB_NumberCreate: 13,
-  
-  WEB_UserInstanceStart: 14,
-  WEB_PostMessage2: 15,
-  WEB_PostMessage3: 16,
-  WEB_PostMessage4: 17,
-  WEB_PostMessage5: 18,
-  WEB_PostMessage6: 19,
-  
-  WEB_WorkerInstance: 20,
-  WEB_MemoryCopy: 21,
-  WEB_MemorySize: 22,
-  WEB_MemoryGrow: 23,
-  WEB_MemoryStrnlen: 24,
-  WEB_MemoryStrncpy: 25,
-  WEB_MemoryFill: 26,
-  WEB_StringCreate: 27,
-  WEB_StringCreateByArgs: 28,
-  WEB_CallObjectMethod: 29,
-  WEB_CallIntegerMethod: 30,
-  WEB_ObjectCreate: 31,
-  WEB_GetObjectProperty: 32,
-  WEB_GetIntegerProperty: 33,
-  WEB_SetObjectProperty: 34,
-  WEB_SetIntegerProperty: 35,
-  WEB_AddEventListener: 36,
-  WEB_PromiseThen: 37,
-  WEB_SetTimeout: 38,
-  WEB_SetInterval: 39,
-  WEB_ClearTimeout: 40,
-  WEB_ClearInterval: 41,
-  WEB_ObjectRetain: 42,
-  WEB_RemoveEventListener: 43,
-  
-  WEB_WorkerPerform2: 44,
-  WEB_WorkerPerform3: 45,
-  WEB_WorkerPerform4: 46,
-  WEB_WorkerPerform5: 47,
-};
